refactor(overlays): move legacy implementation into OverlayLegacy
